:recycle: move particle.py and square.py
