+++ ft_gomoku/game/screen/__init__.py
@@ -1,7 +1,7 @@
 from .title_screen import main_menu
 from .game import game_screen
 from .tutorial import tutorial_screen
-from .particle import Particle, stars_effect
+from ft_gomoku.game.screen.components.particle import Particle, stars_effect
 from .components import draw_rocks
-from .square import Square
+from ft_gomoku.game.screen.components.square import Square
