+++ ft_gomoku/game/screen/components/particle.py
