+++ ft_gomoku/game/screen/components/square.py
